requirements.txt
streamlit==1.16.0
pandas==1.3.5
Pillow==9.3.0